// src/gen/api/types/team_member_permissions_response.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * UserManagement.API
 * OpenAPI spec version: 1.0
 */
import type { ClicksiDataContractsCommonEnumsTeamRole } from "./clicksi_data_contracts_common_enums_team_role";
import type { UserManagementDomainEnumsGeneralTeamPermissions } from "./user_management_domain_enums_general_team_permissions";
import type { UserManagementDomainEnumsProductPermissions } from "./user_management_domain_enums_product_permissions";
import type { UserManagementDomainEnumsFinancialPermissions } from "./user_management_domain_enums_financial_permissions";
import type { UserManagementDomainEnumsAffiliatePermissions } from "./user_management_domain_enums_affiliate_permissions";

export interface TeamMemberPermissionsResponse {
  userId: string;
  teamId: string;
  teamRole: ClicksiDataContractsCommonEnumsTeamRole;
  teamMembershipVersion: number;
  generalPermissions: UserManagementDomainEnumsGeneralTeamPermissions;
  productPermissions: UserManagementDomainEnumsProductPermissions;
  financialPermissions: UserManagementDomainEnumsFinancialPermissions;
  affiliatePermissions: UserManagementDomainEnumsAffiliatePermissions;
}